Add RecetaService, RecetaServiceImpl, and SucursalService classes
- Implement RecetaService with methods for managing recipes, including creating, updating, and retrieving recipes by various parameters.
- Introduce RecetaServiceImpl to provide the implementation of the RecetaService interface, utilizing repositories for data access.
- Add SucursalService for managing branches, including methods for creating, updating, and deleting branches, as well as assigning users to branches and retrieving branches by owner.

+++ backend/src/main/java/com/api/apos/entity/InventarioItem.java
@@ -1,6 +1,5 @@
 package com.api.apos.entity;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Entity;
@@ -27,10 +26,10 @@ public class InventarioItem {
     @JoinColumn(name = "material_id")
     private Material material;
     
-    private BigDecimal cantidad;
-    private BigDecimal stockMinimo;
-    private BigDecimal stockMaximo;
-    private BigDecimal precioUnitario;
+    private Double cantidad;
+    private Double stockMinimo;
+    private Double stockMaximo;
+    private Double precioUnitario;
     
     private LocalDateTime fechaUltimaActualizacion;
     private LocalDateTime fechaUltimaCompra;
